feat: add modificate loader
