+++ news-JS/package.json
@@ -32,5 +32,8 @@
         "webpack-cli": "^5.1.4",
         "webpack-dev-server": "^5.0.0",
         "webpack-merge": "^5.10.0"
+    },
+    "dependencies": {
+        "node-fetch": "^3.3.2"
     }
 }
